+++ src/main.rs
@@ -2,10 +2,11 @@ extern crate eigenvalues;
 extern crate nalgebra as na;
 
 use eigenvalues::davidson::EigenDavidson;
+use eigenvalues::SpectrumTarget;
 
 fn main() {
-  let brr = eigenvalues::utils::generate_diagonal_dominant(10, 0.05);
-  let eig = EigenDavidson::new (brr, 2, "GJD", None).unwrap();
-  println !("eigenvalues:{}", eig.eigenvalues);
-  println !("eigenvectors:{}", eig.eigenvectors);
+    let brr = eigenvalues::utils::generate_diagonal_dominant(10, 0.05);
+    let eig = EigenDavidson::new(brr, 2, "GJD", SpectrumTarget::Lowest).unwrap();
+    println!("eigenvalues:{}", eig.eigenvalues);
+    println!("eigenvectors:{}", eig.eigenvectors);
 }